Updates for 2018.
Starting this year, GCJ no longer gives you the files to download and compute solutions locally. Instead, you submit your code, and they run it on their computers. Thus I had to change some things about input/output to make this work - need to read from stdin and write the solution to stdout instead of saving to a file. Also, I don't think pandas/numpy/etc. are usable anymore.

To make it simpler to go from functions in a notebook to the file to submit, I've written a template file and a function to fill it with your solution function(s) as well as to check that the solution file works properly on the example inputs/outputs (which you have to have already saved appropriately; I've made the function for that nicer too).